+++ R/data.R
@@ -24,15 +24,58 @@
 #' }
 "example_data_5_subject"
 
-#' @title Example CGM mealtime data
+#' @title Example data from Hall et al. (2018)
+#'
+#' @description Dexcom G4 CGM measurements for 19 subjects from the Hall
+#' publicly available dataset. Chosen as a subset of all subjects to be only those
+#' with diabetes or pre-diabetes. Primarily intended for use with example_meals_hall
+#'
+#'
+#' @format a data.frame with 34890 rows and 4 columns, which are:
+#' \describe{
+#' \item{id}{identifier of subject}
+#' \item{time}{date and time stamp}
+#' \item{gl}{glucose level as measured by CGM (mg/dL)}
+#' \item{diagnosis}{character indicating diabetes diagnosis: diabetic or pre-diabetic}
+#' }
+#'
+#' @details
+#' This dataset can be used along with the example_meals_hall dataset in this package
+#' to calculate meal_metrics.
+#'
+#'
+#' @references
+#' Hall et al. (2018) : Glucotypes reveal new patterns of glucose dysregulation
+#' \emph{Plos Biology} \strong{16} (7): 3:e2005143
+#' \doi{10.1371/journal.pbio.2005143}.
+#'
+
+'example_data_hall'
+
+#' @title Example mealtimes data from Hall et al. (2018)
 #'
 #' @description Example of mealtimes data format for meal_metrics function,
-#' corresponds to example_data_1_subject data.
+#' corresponds to example_data_hall data.
 #'
-#' @format A data.frame with 1 row and 3 columns, which are:
+#' @format A data.frame with 9 rows and 3 columns, which are:
 #' \describe{
 #' \item{id}{identifier of subject}
 #' \item{meal}{meal type identifier}
 #' \item{mealtime}{time of meal}
 #' }
-"example_meal"
+#'
+#' @details
+#' There are 3 types of meals available: Cereal Flakes (CF), Peanut Butter
+#' Sandwich (PB), and Protein Bar (Bar). The number after the abbreviation
+#' refers to the replication number for the original study. For more details
+#' on nutritional differences, please see the original study reference.
+#'
+#' This dataset should be used along with example_data_hall to calculate meal_metrics.
+#'
+#'
+#' @references
+#' Hall et al. (2018) : Glucotypes reveal new patterns of glucose dysregulation
+#' \emph{Plos Biology} \strong{16} (7): 3:e2005143
+#' \doi{10.1371/journal.pbio.2005143}.
+
+"example_meals_hall"