+++ R/mage.R
@@ -4,7 +4,7 @@
 #'
 #' @param data DataFrame object with column names "id", "time", and "gl" OR numeric vector of glucose values.
 #'
-#' @param version Either \code{'ma'} or \code{'naive'}. Chooses which version of the MAGE algorithm to use. \code{'ma'} algorithm is more accurate, and is the default. Earlier versions of iglu package (<=2.0.0) used \code{'naive'}.
+#' @param version Either \code{'ma'} or \code{'naive'}. \strong{Default: 'ma'.} Chooses which version of the MAGE algorithm to use. \code{'ma'} algorithm is more accurate, and is the default. Earlier versions of iglu package (<=2.0.0) used \code{'naive'}.
 #'
 #' @param sd_multiplier A numeric value that can change the sd value used to determine size of
 #' glycemic excursions used in the calculation. This is the only parameter that
@@ -14,13 +14,16 @@
 #'
 #' @return A tibble object with two columns: the subject id and corresponding MAGE value.
 #' If a vector of glucose values is passed, then a tibble object with just the MAGE value
-#' is returned. In \code{version = "ma"}, if \code{plot = TRUE}, a list of ggplots will
-#' be returned with one plot per subject.
+#' is returned.
+#'
+#' In \code{version = "ma"}, if \code{plot = TRUE}, a list of ggplots will
+#' be returned with one plot per subject. To return an interactive plot, use iglu::mage_ma_single with \code{plot_type='plotly'} on each subject individually.
+#'
 #' @export
 #'
 #' @details If version \code{'ma'} is selected, the function computationally emulates the manual method for calculating the mean amplitude of glycemic excursions (MAGE) first suggested in "Mean Amplitude of Glycemic Excursions, a Measure of Diabetic Instability", (Service, 1970). For this version, glucose values will be interpolated over a uniform time grid prior to calculation.
 #'
-#' \code{'ma'} is a more accurate algorithm that uses the crosses of a short and long moving average to identify intervals where a peak/nadir might exist. Then, the height from one peak/nadir to the next nadir/peak is calculated from the *original* (not moving average) glucose values.
+#' \code{'ma'} is a more accurate algorithm that uses the crosses of a short and long moving average to identify intervals where a peak/nadir might exist. Then, the height from one peak/nadir to the next nadir/peak is calculated from the _original_ (not moving average) glucose values.
 #'
 #' \code{'naive'} algorithm calculates MAGE by taking the mean of absolute glucose differences (between each value and the mean) that are greater than the standard deviation. A multiplier can be added to the standard deviation using the \code{sd_multiplier} argument.
 #'
@@ -33,10 +36,15 @@
 #' data(example_data_5_subject)
 #' mage(example_data_5_subject, version = 'ma')
 
-mage <- function(data, version = c('ma', 'naive'), sd_multiplier = 1,
-                 short_ma = 5, long_ma = 32, type = c('auto', 'plus', 'minus'),
-                 plot = FALSE, dt0 = NULL, inter_gap = 45, tz = "",
-                 title = NA, xlab = NA, ylab = NA, show_ma = FALSE) {
+mage <- function(data,
+                 version = c('ma', 'naive'),
+                 sd_multiplier = 1,
+                 short_ma = 5, long_ma = 32,
+                 return_type = c('num', 'df'),
+                 direction = c('avg', 'service', 'max', 'plus', 'minus'),
+                 dt0 = NULL,  tz = "", inter_gap = 45,
+                 max_gap=180,
+                 plot = FALSE, title = NA, xlab = NA, ylab = NA, show_ma = FALSE, show_excursions = TRUE) {
 
   # Match version
   version = match.arg(version)
@@ -46,26 +54,33 @@ mage <- function(data, version = c('ma', 'naive'), sd_multiplier = 1,
     return(mage_sd(data, sd_multiplier = sd_multiplier))
   }
 
-  return(mage_ma(data, short_ma = short_ma, long_ma = long_ma, type = type,
-                 plot = plot, dt0 = dt0, inter_gap = inter_gap, tz = tz,
-                 title = title, xlab = xlab, ylab = ylab, show_ma = show_ma))
+  direction = match.arg(direction, c('avg', 'service', 'max', 'plus', 'minus'))
+
+  return(mage_ma(data, short_ma = short_ma, long_ma = long_ma, return_type=return_type, direction=direction,
+                 plot = plot, dt0 = dt0, inter_gap = inter_gap, max_gap = max_gap, tz = tz,
+                 title = title, xlab = xlab, ylab = ylab, show_ma = show_ma, show_excursions=show_excursions))
 }
 
-mage_ma <- function(data, short_ma = 5, long_ma = 32, type = c('auto', 'plus', 'minus'),
-                    plot = FALSE, dt0 = NULL, inter_gap = 45, tz = "",
-                    title = NA, xlab = NA, ylab = NA, show_ma = FALSE) {
+mage_ma <- function(data,
+                    short_ma = 5, long_ma = 32,
+                    return_type = c('num', 'df'),
+                    direction = c('avg', 'service', 'max', 'plus', 'minus'),
+                    dt0 = NULL, inter_gap = 45, tz = "",
+                    max_gap = 180,
+                    plot = FALSE, title = NA, xlab = NA, ylab = NA, show_ma = FALSE, show_excursions=TRUE) {
   id = . = MAGE = NULL
   rm(list = c("id", ".", "MAGE"))
 
   data = check_data_columns(data)
-  is_vector = attr(data, "is_vector") # TODO: is this check really necessary? when does this return true?
+  is_vector = attr(data, "is_vector")
+  direction = match.arg(direction, c('avg', 'service', 'max', 'plus', 'minus'))
 
   out <- data %>%
     dplyr::filter(!is.na(gl)) %>%
     dplyr::group_by(id) %>%
-    dplyr::do(MAGE = mage_ma_single(., short_ma = short_ma, long_ma = long_ma, type = type,
-                                    plot = plot, dt0 = dt0, inter_gap = inter_gap, tz = tz,
-                                    title = title, xlab = xlab, ylab = ylab, show_ma = show_ma))
+    dplyr::do(MAGE = mage_ma_single(., short_ma = short_ma, long_ma = long_ma, return_type=return_type, direction=direction,
+                                    plot = plot, dt0 = dt0, inter_gap = inter_gap, max_gap = max_gap, tz = tz,
+                                    title = title, xlab = xlab, ylab = ylab, show_ma = show_ma, show_excursions = show_excursions))
 
   # Check if a ggplot or number in list is returned - convert the latter to a number
   if(class(out$MAGE[[1]])[1] == "numeric" | is.na(out$MAGE[[1]][1])) {