+++ NEWS.md
@@ -1,3 +1,26 @@
+# iglu 4.0.0
+* CRAN release to match upcoming paper summarizing updated MAGE algorithm, episode calculation functionality, meal metrics, clustering and new example data inclusion
+
+
+# iglu 3.5.3
+* minor bugfix in MAGE: stopped extraneous warning from being thrown
+* Added tests for mage + mean + median + sd
+
+# iglu 3.5.2
+* MAGE plot can now show excursions via "show_excursions = TRUE" flag
+* debugged formatting error with Mage vignette
+* debugged internal build warnings
+* Metrics heatmap removes metrics with no variability across subjects
+* Add example dataset to demonstrate meal metrics calculations
+
+# iglu 3.5.1
+* MAGE handles traces with gaps, restarting calculation when gap >= user-defined threshold
+* MAGE returns ggplot as well as plotly
+* Fixed plotting functionality with large gaps and multiple segments
+* MAGE "left-side accumulation" optimization
+* Added "MAGE_max" and "MAGE_avg" to possible directions able to be calculated (in addition to MAGE+, MAGE-, and Service's definition)
+* Updated MAGE vignette for more details
+
 # iglu 3.5.0
 * minor bug corrections in mage calculations
 * adjustment from summarize to reframe
@@ -7,7 +30,7 @@
 * add AGP plot smoothing
 * add functionality to create a heatmap of metrics by subject
 
-#iglu 3.4.3
+# iglu 3.4.3
 * bug correction for denominator in hypo_index and hyper_index in case of missing glucose values
 
 # iglu 3.4.2